server/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
distlib==0.4.0
Django==4.2.26
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
filelock==3.19.1
packaging==25.0
pipenv==2025.0.4
platformdirs==4.4.0
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.15.0
virtualenv==20.35.4
